load route transitions by route, instead of category
Before we loaded async requirements by category, in parallel:

- parallel
  - modules
    - route1 module
    - route2 module
  - styles
    - route1 styles
    - route2 styles
  - data
    - route1 data
    - route2 data

This commit flips it around and loads by routes all in parallel instead of categories in parallel

- parallel
  - route1
    - module
    - styles
    - data
  - route 2
    - module
    - styles
    - data

This change is anticipating the `links` export. We’ll need the route data and the route module before we can start loading the links (need the module to even have the link function, need the data to pass to it so they can preload based on data).

This would create unnecessary interdependencies between one route’s data blocking another route’s links. If we kept loading by category instead of route, the best we could do is this:

- parallel
  - all route modules
  - all route data
- parallel
  - all route links

This means that if route1’s module and data is ready, we can’t start loading their links until route2’s data is ready. Loading by route we can, and it will look like this:


- parallel
  - route1
    - parallel
      - module
      - data
    - link
  - route2
    - parallel
      - module
      - data
    - link

This way if route2 has loaded its module and data it can immediately start loading its link assets even while route1 is still loading data.