chore(docs): update tutorial to use db variable for continuity (#1602)
* update tutorial to use db varaible for continuity

Earlier in the tutorial the pattern have been using `db` in stead of `prisma`. When `prisma` suddenly is used in the seed and only the highlighted diff is copied to the editor, the seed function will fail as `prisma` is not defined. However, `db` is if the user have copied `seed.ts` from earlier in the tutorial.

* Update contributors.yml