feat(compiler): officially support Deno (#3117)
* docs(decisions): use `npm` to manage NPM dependencies for Deno projects

* refactor(compiler): do not bundle url imports

keep url imports as-is in build output so that each runtime can handle url imports in their own way.
for example, deno will download dependencies from url imports at runtime. the browser will download
dependendencies from url imports on-demand. node will simply throw an error when encountering a url
import.

since we will recommend to use npm/node_modules for dependency
management (even for deno), the only url imports we expect are for deno
imports on the server from https://deno.land or from Deno-friendly CDNs
like https://esm.sh .

* feat(compiler): can build deno-compatible server bundle

* feat(@remix-run/deno): add server runtime package for deno

* feat(templates/deno): use npm-managed dependencies

* test(replace-remix-imports): fix `shell.grep` call

previously, `shell.grep` was being called with the `-r` option, which it does not recognize,
resulting in an exit ode of 1. we incorrectly interpreted that as `grep` finishing the search
without finding matches. instead, the `shell.grep` call was crashing. now we correctly pass a list
of files to `shell.grep` without the `-r` option.

* test(compiler): browser and server bundles for deno

* refactor(@remix-run/deno): throw custom error when static files are not found

so that `createRequestHandlerWithStaticFiles` is more portable across runtimes and adapters