fix(remix-dev): Fix private stack validation (#2804)
Include `GITHUB_TOKEN` when validating templates. This is necessary in order to create a stack from a private GitHub repo. Fixes #2801.

It looks like `GITHUB_TOKEN` is ignored for web requests to private repos: even when that token should grant access, the server returns a 404.

Instead, we can hit the API, which is happy to accept a token. This works for public repos as well, with or without a `GITHUB_TOKEN`. However, we do need to do a bit of extra work when a `filePath` is passed. The API response only tells us some metadata, we need the `/tree` call to learn about what files exist there.

Co-authored-by: Logan McAnsh <logan@mcan.sh>